Extract NodeHandleCommand
BrushVertexCommand's reason d'etre was to update the handle manager when the
corresponding nodes were changed by removing the handles being changed, then
readding and reselecting them in their new positions. This concept is also
generlaized to the new node handle concept.
