+++ lib/TbMdlLib/CMakeLists.txt
@@ -16,7 +16,6 @@ target_sources(TbMdlLib
     ${CMAKE_CURRENT_SOURCE_DIR}/src/BrushFaceReference.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/src/BrushNode.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/src/BrushRendererBrushCache.cpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/src/BrushVertexCommands.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/src/CircleShape.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/src/ColorRange.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/src/Command.cpp
